+++ CHANGES.md
@@ -142,6 +142,7 @@ UI:
 -   Added tooltips to the `Production` section of the `People and Production` page
 -   Reworded the user access options
 -   Removed help icons without content
+-   Improve keywords generation logic for Spreadsheet
 
 Gateway: